update the existing file

+++ gcr-codebase/programming-prequisites/CircleArea.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class CircleArea{
   public static void main(String[] args ){
      Scanner sc = new Scanner(System.in);
-	 float r = sc.nextFloat();
+	 double r = sc.nextDouble();
 	 
-	 float area = 3.14 * r*r;
+	 double area = 3.14 * r*r;
 	 
 	 System.out.println(area);
 	}
